backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
Pillow==10.3.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
pytest==8.2.0
pytest-django==4.8.0
requests==2.31.0
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
